+++ src/InternalsVisibleTo.cs
@@ -1,4 +1,5 @@
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("SharpImageConverter.Tests")]
+[assembly: InternalsVisibleTo("SharpImageConverter.Cli")]